fix: class moved common folder
