+++ nightfall-optimist/src/classes/index.mjs
@@ -1,0 +1,21 @@
+import BlockError from './block-error.mjs';
+import Fq2 from './fq2.mjs';
+import Proposer from './proposer.mjs';
+import TransactionError from './transaction-error.mjs';
+import VerificationKey from './verification-key.mjs';
+
+// below file are common file and volume shared via docker-compose.yml
+import Proof from './proof.mjs'; // eslint-disable-line import/no-unresolved
+import PublicInputs from './public-inputs.mjs'; // eslint-disable-line import/no-unresolved
+import Transaction from './transaction.mjs'; // eslint-disable-line import/no-unresolved
+
+export {
+  Proof,
+  PublicInputs,
+  Transaction,
+  BlockError,
+  Fq2,
+  Proposer,
+  TransactionError,
+  VerificationKey,
+};
